Revert "Refactoring project to use spring boot, allowing annotation of class …"
