+++ CarDriver.java
@@ -1,0 +1,5 @@
+public class CarDriver extends User{
+    public CarDriver(int userID) {
+        super(userID);
+    }
+}
