+++ User.java
@@ -1,0 +1,12 @@
+public class User {
+    private final int userID;
+    private final String userFirstName;
+    private final String userLastName;
+
+
+    public User(int userID, String userFirstName, String userLastName) {
+        this.userID = userID;
+        this.userFirstName = userFirstName;
+        this.userLastName = userLastName;
+    }
+}
