+++ Vehicle.java
@@ -1,5 +1,3 @@
-package com.RentFast.App.Entities;
-
 public class Vehicle {
 
     private final String licensePlate;
