--- src/com/RentFast/App/Billing/InsuranceController.java
@@ -1,4 +1,0 @@
-package com.RentFast.App.Billing;
-
-public class InsuranceController {
-}
