--- src/com/RentFast/App/Billing/PromotionController.java
@@ -1,4 +1,0 @@
-package com.RentFast.App.Billing;
-
-public class PromotionController {
-}
