--- src/com/RentFast/App/Billing/SubscriptionController.java
@@ -1,4 +1,0 @@
-package com.RentFast.App.Billing;
-
-public class SubscriptionController {
-}
