--- src/com/RentFast/App/Entities/CarDriver.java
@@ -1,8 +1,0 @@
-package com.RentFast.App.Entities;
-
-public class CarDriver extends User{
-
-    public CarDriver(int userID, String userFirstName, String userLastName, int phoneNumber, String licenseId, long signUpDate) {
-        super(userID, userFirstName, userLastName, phoneNumber, licenseId, signUpDate);
-    }
-}
